Tests: Configure Capybara default_driver rather than current_driver
It seems the intent here was to use selenium_chrome_headless for all
Cucumber tests.  But the default_driver value would still have been
rack_test -- which causes many test failures when we get to Rails 6.
Rather than debugging the rack_test failures, I'm setting the
default_driver to what appears to have been the original intent.

rack_test has a different interpretation of `visible?` than selenium, as
selenium requires some content in the div.  So I've added some content to
an empty div in a test to get it to pass.  This change only affects the
test suite for capybara-ui and not any consumers of the capybara-ui
library.